refactor + fix env-based config loading
