+++ metadata-ingestion/src/datahub/emitter/rest_emitter.py
@@ -10,11 +10,8 @@ from deprecated import deprecated
 from requests.adapters import HTTPAdapter, Retry
 from requests.exceptions import HTTPError, RequestException
 
-from datahub.cli.cli_utils import (
-    ensure_has_system_metadata,
-    fixup_gms_url,
-    get_system_auth,
-)
+from datahub.cli.cli_utils import ensure_has_system_metadata, fixup_gms_url
+from datahub.cli.config_utils import get_system_auth
 from datahub.configuration.common import ConfigurationError, OperationalError
 from datahub.emitter.generic_emitter import Emitter
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
