+++ metadata-ingestion/src/datahub/entrypoints.py
@@ -13,7 +13,11 @@ from datahub.cli.cli_utils import (
     generate_access_token,
     make_shim_command,
 )
-from datahub.cli.config_utils import DATAHUB_CONFIG_PATH, get_boolean_env_variable
+from datahub.cli.config_utils import (
+    DATAHUB_CONFIG_PATH,
+    get_boolean_env_variable,
+    write_gms_config,
+)
 from datahub.cli.delete_cli import delete
 from datahub.cli.docker_cli import docker
 from datahub.cli.exists_cli import exists
@@ -33,7 +37,7 @@ from datahub.cli.state_cli import state
 from datahub.cli.telemetry import telemetry as telemetry_cli
 from datahub.cli.timeline_cli import timeline
 from datahub.configuration.common import should_show_stack_trace
-from datahub.ingestion.graph.client import get_default_graph, write_gms_config
+from datahub.ingestion.graph.client import get_default_graph
 from datahub.telemetry import telemetry
 from datahub.utilities._custom_package_loader import model_version_name
 from datahub.utilities.logging_manager import configure_logging
@@ -149,7 +153,7 @@ def init(use_password: bool = False) -> None:
             type=str,
             default="",
         )
-    write_gms_config(host, token)
+    write_gms_config(host, token, merge_with_previous=False)
 
     click.echo(f"Written to {DATAHUB_CONFIG_PATH}")
