+++ metadata-ingestion/src/datahub/ingestion/graph/client.py
@@ -3,8 +3,6 @@ import enum
 import functools
 import json
 import logging
-import os
-import sys
 import textwrap
 import time
 from dataclasses import dataclass
@@ -24,19 +22,19 @@ from typing import (
     Union,
 )
 
-import click
 from avro.schema import RecordSchema
 from deprecated import deprecated
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
 from requests.models import HTTPError
 
-from datahub.cli import config_utils
+from datahub.cli.config_utils import load_client_config
 from datahub.configuration.common import ConfigModel, GraphError, OperationalError
 from datahub.emitter.aspect import TIMESERIES_ASPECT_MAP
 from datahub.emitter.mce_builder import DEFAULT_ENV, Aspect
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
 from datahub.emitter.rest_emitter import DatahubRestEmitter
 from datahub.emitter.serialization_helper import post_json_transform
+from datahub.ingestion.graph.config import DatahubClientConfig
 from datahub.ingestion.graph.connections import (
     connections_gql,
     get_id_from_connection_urn,
@@ -89,26 +87,6 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 _MISSING_SERVER_ID = "missing"
 _GRAPH_DUMMY_RUN_ID = "__datahub-graph-client"
-
-ENV_METADATA_HOST_URL = "DATAHUB_GMS_URL"
-ENV_METADATA_TOKEN = "DATAHUB_GMS_TOKEN"
-ENV_METADATA_HOST = "DATAHUB_GMS_HOST"
-ENV_METADATA_PORT = "DATAHUB_GMS_PORT"
-ENV_METADATA_PROTOCOL = "DATAHUB_GMS_PROTOCOL"
-
-
-class DatahubClientConfig(ConfigModel):
-    """Configuration class for holding connectivity to datahub gms"""
-
-    server: str = "http://localhost:8080"
-    token: Optional[str] = None
-    timeout_sec: Optional[int] = None
-    retry_status_codes: Optional[List[int]] = None
-    retry_max_times: Optional[int] = None
-    extra_headers: Optional[Dict[str, str]] = None
-    ca_certificate_path: Optional[str] = None
-    client_certificate_path: Optional[str] = None
-    disable_ssl_verification: bool = False
 
 
 # Alias for backwards compatibility.
@@ -1779,84 +1757,3 @@ def get_default_graph() -> DataHubGraph:
     graph = DataHubGraph(graph_config)
     graph.test_connection()
     return graph
-
-
-class DatahubConfig(BaseModel):
-    gms: DatahubClientConfig
-
-
-config_override: Dict = {}
-
-
-def get_details_from_env() -> Tuple[Optional[str], Optional[str]]:
-    host = os.environ.get(ENV_METADATA_HOST)
-    port = os.environ.get(ENV_METADATA_PORT)
-    token = os.environ.get(ENV_METADATA_TOKEN)
-    protocol = os.environ.get(ENV_METADATA_PROTOCOL, "http")
-    url = os.environ.get(ENV_METADATA_HOST_URL)
-    if port is not None:
-        url = f"{protocol}://{host}:{port}"
-        return url, token
-    # The reason for using host as URL is backward compatibility
-    # If port is not being used we assume someone is using host env var as URL
-    if url is None and host is not None:
-        logger.warning(
-            f"Do not use {ENV_METADATA_HOST} as URL. Use {ENV_METADATA_HOST_URL} instead"
-        )
-    return url or host, token
-
-
-def load_client_config() -> DatahubClientConfig:
-    try:
-        ensure_datahub_config()
-        client_config_dict = config_utils.get_client_config()
-        datahub_config: DatahubClientConfig = DatahubConfig.parse_obj(
-            client_config_dict
-        ).gms
-    except ValidationError as e:
-        click.echo(
-            f"Received error, please check your {config_utils.CONDENSED_DATAHUB_CONFIG_PATH}"
-        )
-        click.echo(e, err=True)
-        sys.exit(1)
-
-    # Override gms & token configs if specified.
-    if len(config_override.keys()) > 0:
-        datahub_config.server = str(config_override.get(ENV_METADATA_HOST_URL))
-        datahub_config.token = config_override.get(ENV_METADATA_TOKEN)
-    elif config_utils.should_skip_config():
-        gms_host_env, gms_token_env = get_details_from_env()
-        if gms_host_env:
-            datahub_config.server = gms_host_env
-        datahub_config.token = gms_token_env
-
-    return datahub_config
-
-
-def ensure_datahub_config() -> None:
-    if not os.path.isfile(config_utils.DATAHUB_CONFIG_PATH):
-        click.secho(
-            f"No {config_utils.CONDENSED_DATAHUB_CONFIG_PATH} file found, generating one for you...",
-            bold=True,
-        )
-        write_gms_config(config_utils.DEFAULT_GMS_HOST, None)
-
-
-def write_gms_config(
-    host: str, token: Optional[str], merge_with_previous: bool = True
-) -> None:
-    config = DatahubConfig(gms=DatahubClientConfig(server=host, token=token))
-    if merge_with_previous:
-        try:
-            previous_config = config_utils.get_client_config()
-            assert isinstance(previous_config, dict)
-        except Exception as e:
-            # ok to fail on this
-            previous_config = {}
-            logger.debug(
-                f"Failed to retrieve config from file {config_utils.DATAHUB_CONFIG_PATH}: {e}. This isn't fatal."
-            )
-        config_dict = {**previous_config, **config.dict()}
-    else:
-        config_dict = config.dict()
-    config_utils.persist_datahub_config(config_dict)
